requirements.txt
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
beautifulsoup4==4.7.1
colorama==0.4.1
coverage==4.5.3
dj-database-url==0.5.0
Django==2.1.8
django-autocomplete-light==3.3.5
django-braces==1.13.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-extensions==2.1.9
django-filter==2.1.0
django-fsm==2.6.1
django-guardian==2.0.0
django-material==1.5.6
django-pyodbc-azure==2.1.0.0
django-tables2==2.0.6
django-viewflow==1.5.6
django-webtest==1.9.6
factory-boy==2.12.0
Faker==1.0.7
graphviz==0.10.1
gunicorn==19.9.0
importlib-metadata==0.17
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
more-itertools==7.0.0
packaging==19.0
pluggy==0.12.0
py==1.8.0
pycodestyle==2.5.0
pylint==2.3.1
pyodbc==4.0.26
pyparsing==2.4.0
pytest==4.6.1
pytest-django==3.5.0
python-dateutil==2.8.0
python-dotenv==0.10.2
pytz==2019.1
selenium==3.141.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
text-unidecode==1.2
typed-ast==1.3.5
urllib3==1.25.3
waitress==1.3.0
wcwidth==0.1.7
WebOb==1.8.5
WebTest==2.0.33
Werkzeug==0.15.4
whitenoise==4.1.2
wrapt==1.11.1
zipp==0.5.1
